@@ -8,72 +8,82 @@
     "dev": "webpack serve --mode development --progress --config build/webpack.dev.conf.js",
     "start": "npm run dev",
     "build": "node build/build.js",
-    "electron": "cross-env NODE_ENV=dev node_modules/.bin/electron pack/electron",
+    "electron": "cross-env ARDM_ENV=development electron pack/electron",
     "lint": "eslint --ext js,vue src --fix",
     "lint:init": "eslint --init",
-    "asar": "node_modules/.bin/asar pack dist dist/app.asar --unpack-dir \"{app.asar,build-apps,app.asar.unpacked}\"",
+    "asar": "asar pack dist dist/app.asar --unpack-dir \"{app.asar,build-apps,app.asar.unpacked}\"",
     "pack:prepare": "npm run build && cp -r pack/electron/* dist/",
-    "pack:win": "node_modules/.bin/electron-builder --project=dist -w -p never",
-    "pack:win32": "node_modules/.bin/electron-builder --project=dist -w --ia32 -p never",
-    "pack:mac": "node_modules/.bin/electron-builder --project=dist -m -p never",
-    "pack:linux": "node_modules/.bin/electron-builder --project=dist -l -p never",
-    "pack:win:publish": "node_modules/.bin/electron-builder --project=dist -w -p always",
-    "pack:mac:publish": "node_modules/.bin/electron-builder --project=dist -m -p always",
-    "pack:macm1:publish": "node_modules/.bin/electron-builder --project=dist -m --arm64 -p always -c.artifactName='${productName} M1 ${arch} ${version}.${ext}'",
-    "pack:linux:publish": "node_modules/.bin/electron-builder --project=dist -l -p always",
-    "pack:win:docker": "docker run --rm -it --env GH_TOKEN=$GH_TOKEN --env ELECTRON_CACHE=\"/root/.cache/electron\" --env ELECTRON_BUILDER_CACHE=\"/root/.cache/electron-builder\" -v ${PWD}:/project -v ~/.cache/electron:/root/.cache/electron -v ~/.cache/electron-builder:/root/.cache/electron-builder electronuserland/builder:wine sh -c \"node_modules/.bin/electron-builder --project=dist -w -p never\"",
-    "et": "node_modules/.bin/et -m",
+    "pack:win": "electron-builder --project=dist -w -p never",
+    "pack:win32": "electron-builder --project=dist -w --ia32 -p never",
+    "pack:mac": "electron-builder --project=dist -m -p never",
+    "pack:mas": "CSC_IDENTITY_AUTO_DISCOVERY=false npm run pack:mac",
+    "pack:linux": "electron-builder --project=dist -l -p never",
+    "pack:win:publish": "electron-builder --project=dist -w -p always",
+    "pack:mac:publish": "electron-builder --project=dist -m -p always",
+    "pack:macm1:publish": "electron-builder --project=dist -m --arm64 -p always -c.artifactName='${productName} M1 ${arch} ${version}.${ext}'",
+    "pack:linux:publish": "electron-builder --project=dist -l -p always",
+    "pack:win:docker": "docker run --rm -it --env GH_TOKEN=$GH_TOKEN --env ELECTRON_CACHE=\"/root/.cache/electron\" --env ELECTRON_BUILDER_CACHE=\"/root/.cache/electron-builder\" -v ${PWD}:/project -v ~/.cache/electron:/root/.cache/electron -v ~/.cache/electron-builder:/root/.cache/electron-builder electronuserland/builder:wine sh -c \"electron-builder --project=dist -w -p never\"",
+    "et": "et -m",
     "sign": "cd pack/mas; bash sign.sh"
   },
   "dependencies": {
-    "@msgpack/msgpack": "^2.3.0",
-    "@ztree/ztree_v3": "^3.5.46",
-    "brotli": "^1.3.2",
+    "@qii404/ioredis": "^4.28.501",
+    "@qii404/json-bigint": "^1.0.0",
+    "@qii404/redis-splitargs": "^1.0.1",
+    "@qii404/vue-easy-tree": "^1.0.10",
+    "algo-msgpack-with-bigint": "^2.1.1",
     "element-ui": "^2.4.11",
     "font-awesome": "^4.7.0",
-    "ioredis": "^4.27.0",
-    "json-bigint": "git+https://github.com/qishibo/json-bigint.git",
     "keymaster": "^1.6.2",
     "monaco-editor": "^0.30.1",
     "php-serialize": "^4.0.2",
-    "redis-splitargs": "git+https://github.com/qishibo/splitargs.git",
+    "protobufjs": "^6.11.2",
+    "rawproto": "^0.7.6",
     "sortablejs": "^1.14.0",
-    "tunnel-ssh": "^4.1.4",
-    "vue": "^2.5.2",
-    "vue-i18n": "^8.7.0"
+    "tunnel-ssh": "^4.1.6",
+    "vue": "^2.6.11",
+    "vue-i18n": "^8.7.0",
+    "vue-virtual-scroller": "^1.1.2"
   },
   "devDependencies": {
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.0.0",
+    "@babel/plugin-proposal-decorators": "^7.0.0",
+    "@babel/plugin-proposal-export-namespace-from": "^7.0.0",
+    "@babel/plugin-proposal-function-sent": "^7.0.0",
+    "@babel/plugin-proposal-json-strings": "^7.0.0",
+    "@babel/plugin-proposal-numeric-separator": "^7.0.0",
+    "@babel/plugin-proposal-throw-expressions": "^7.0.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
+    "@babel/plugin-syntax-import-meta": "^7.0.0",
+    "@babel/plugin-syntax-jsx": "^7.0.0",
+    "@babel/plugin-transform-runtime": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@vue/babel-preset-jsx": "^1.2.4",
     "asar": "^1.0.0",
     "autoprefixer": "^7.1.2",
-    "babel-core": "^6.22.1",
     "babel-helper-vue-jsx-merge-props": "^2.0.3",
-    "babel-loader": "^7.1.1",
+    "babel-loader": "^8.0.0",
     "babel-plugin-component": "^1.1.1",
-    "babel-plugin-syntax-jsx": "^6.18.0",
-    "babel-plugin-transform-runtime": "^6.22.0",
-    "babel-plugin-transform-vue-jsx": "^3.5.0",
-    "babel-preset-env": "^1.3.2",
-    "babel-preset-stage-2": "^6.22.0",
     "chalk": "^2.0.1",
     "copy-webpack-plugin": "^4.6.0",
     "cross-env": "^5.2.0",
     "css-loader": "^0.28.0",
-    "electron": "^11.4.9",
-    "electron-builder": "^22.10.5",
-    "element-theme": "^2.0.1",
+    "electron": "^12.2.3",
+    "electron-builder": "^22.14.5",
     "element-theme-chalk": "^2.13.0",
     "eslint": "^5.14.1",
     "eslint-config-airbnb-base": "^13.1.0",
     "eslint-plugin-import": "^2.16.0",
     "eslint-plugin-vue": "^5.2.2",
-    "extract-text-webpack-plugin": "^3.0.0",
     "file-loader": "^1.1.4",
-    "font-list": "^1.2.11",
+    "font-list": "^1.4.5",
     "friendly-errors-webpack-plugin": "^1.6.1",
     "html-webpack-plugin": "^4.5.2",
     "js-yaml": ">=3.13.1",
     "mini-css-extract-plugin": "^0.11.3",
     "monaco-editor-webpack-plugin": "^6.0.0",
+    "node-loader": "^1.0.3",
     "node-notifier": "^5.1.2",
     "optimize-css-assets-webpack-plugin": "^5.0.3",
     "ora": "^1.2.0",
