@@ -1,18 +1,42 @@
 // Modules to control application life and create native browser window
-const { app, BrowserWindow, Menu, ipcMain } = require('electron');
+const { app, BrowserWindow, Menu, ipcMain, dialog } = require('electron');
 const fontManager = require('./font-manager');
 const winState = require('./win-state');
 
+const url = require('url');
+const path = require('path');
 
-global.APP_ENV = (process.env.NODE_ENV === 'dev') ? 'dev' : 'production';
+// disable GPU for some white screen issues
+// app.disableHardwareAcceleration();
+// app.commandLine.appendSwitch('disable-gpu');
 
-if (APP_ENV === 'production') {
-  require('./update')();
-}
+global.APP_ENV = (process.env.ARDM_ENV === 'development') ? 'development' : 'production';
 
 // Keep a global reference of the window object, if you don't, the window will
 // be closed automatically when the JavaScript object is garbage collected.
 let mainWindow;
+
+// handle uncaught exception
+process.on('uncaughtException', (err, origin) => {
+  if (!err) {
+    return;
+  }
+
+  dialog.showMessageBoxSync(mainWindow, {
+    type: 'error',
+    title: 'Whoops! Uncaught Exception', 
+    message: err.stack,
+    detail: '\nDon\'t worry, I will fix it! 😎😎\n\n'
+            + 'Submit issue to: \nhttps://github.com/qishibo/AnotherRedisDesktopManager/'
+  });
+
+  process.exit();
+});
+
+// auto update
+if (APP_ENV === 'production') {
+  require('./update')();
+}
 
 function createWindow() {
   // get last win stage
@@ -20,16 +44,17 @@ function createWindow() {
 
   // Create the browser window.
   mainWindow = new BrowserWindow({
-    x: (lastWinStage.x > 0) ? lastWinStage.x : null,
-    y: (lastWinStage.y > 0) ? lastWinStage.y : null,
-    width: (lastWinStage.width > 250) ? lastWinStage.width : 1100,
-    height: (lastWinStage.height > 250) ? lastWinStage.height : 728,
+    x: lastWinStage.x,
+    y: lastWinStage.y,
+    width: lastWinStage.width,
+    height: lastWinStage.height,
     icon: `${__dirname}/icons/icon.png`,
     autoHideMenuBar: true,
     webPreferences: {
       nodeIntegration: true,
       // add this to keep 'remote' module avaiable. Tips: it will be removed in electron 14
       enableRemoteModule: true,
+      contextIsolation: false,
     },
   });
 
@@ -42,13 +67,22 @@ function createWindow() {
   // and load the index.html of the app.
   if (APP_ENV === 'production') {
     // mainWindow.loadFile('index.html');
-    mainWindow.loadURL(`file://${__dirname}/index.html?version=${app.getVersion()}`);
+    mainWindow.loadURL(url.format({
+      protocol: 'file',
+      slashes: true,
+      pathname: path.join(__dirname, 'index.html'),
+      query: {version: app.getVersion()},
+    }));
   } else {
     mainWindow.loadURL(`http://localhost:9988/?version=${app.getVersion()}`);
   }
 
   // Open the DevTools.
   // mainWindow.webContents.openDevTools();
+
+  mainWindow.on('close', () => {
+    mainWindow.webContents.send('closingWindow');
+  });
 
   // Emitted when the window is closed.
   mainWindow.on('closed', () => {
