@@ -1,7 +1,7 @@
 <template>
 <el-dialog @open='openDialog' :title="$t('message.command_log')" :visible.sync="visible" custom-class='command-log-dialog' width="90%" append-to-body>
   <ul ref='commandLogUl' class="command-log-ul">
-    <li v-for="record of logs">
+    <li v-for="record of logsShow">
       <span class="command-time">{{ record.time }} -</span>
       <span class="command-con-name">[{{ record.connectionName }}]:&nbsp;</span>
       <span class="command-name">{{ record.name }} </span>
@@ -9,6 +9,11 @@
       <span class="command-cost">[{{ record.cost }}ms]</span>
     </li>
   </ul>
+
+  <!-- filter -->
+  <el-input v-model='filter' size='mini' style='max-width: 200px;' :placeholder="$t('message.key_to_search')"></el-input>&nbsp;
+  <!-- show only write commands -->
+  <el-checkbox v-model='showOnlyWrite'>Only Write</el-checkbox>
 
   <div slot="footer" class="dialog-footer">
     <el-button @click="logs=[]">{{ $t('el.colorpicker.clear') }}</el-button>
@@ -18,11 +23,16 @@
 </template>
 
 <script type="text/javascript">
+import {writeCMD} from '@/commands.js';
+
 export default {
   data() {
     return {
       visible: false,
       logs: [],
+      maxLength: 2000,
+      filter: '',
+      showOnlyWrite: false,
     };
   },
   created() {
@@ -42,8 +52,28 @@ export default {
         connectionName: record.connectionName,
       });
 
+      this.logs.length > this.maxLength && (this.logs = this.logs.slice(-this.maxLength));
       this.visible && this.scrollToBottom();
     });
+  },
+  computed: {
+    logsShow() {
+      let logs = this.logs;
+
+      if (this.showOnlyWrite) {
+        logs = logs.filter(item => {
+          return writeCMD[item.name.toUpperCase()];
+        });
+      }
+
+      if (this.filter) {
+        logs = logs.filter(item => {
+          return item.name.includes(this.filter) || item.args.includes(this.filter);
+        });
+      }
+
+      return logs;
+    },
   },
   methods: {
     show() {
