@@ -1,62 +1,63 @@
 <template>
 <div>
-  <el-card v-loading="loadingDelete||loadingScan" class="box-card del-batch-card">
+  <el-card class="box-card del-batch-card">
     <!-- card title -->
     <div slot="header" class="clearfix">
-      <span class="del-title">{{ $t('message.keys_to_be_deleted') }}</span>
+      <span class="del-title"><i class="fa fa-exclamation-triangle"></i> {{ $t('message.keys_to_be_deleted') }}</span>
       <i v-if="loadingScan||loadingDelete" class='el-icon-loading'></i>
       <el-tag size="mini">
         <span v-if="loadingScan">Scanning... </span>
         <span v-if="loadingDelete">Deleting... </span>
-        Total: {{ Object.keys(allKeys).length }}
+        Total: {{ allKeysList.length }}
       </el-tag>
 
-      <el-button @click="confirmDelete" :disabled="loadingScan||loadingDelete" style="float: right;" type="danger">{{ $t('message.delete_all') }}</el-button>
+      <!-- del btn -->
+      <el-button @click="confirmDelete" :disabled="loadingScan||loadingDelete||allKeysList.length == 0" style="float: right;" type="danger">{{ $t('message.delete_all') }}</el-button>
+      <!-- toggle scanning btn -->
+      <el-button v-if="rule.pattern.length && !scanningEnd" @click="toggleScanning()" type="text" style="float: right;">{{loadingScan ? $t('message.pause') : $t('message.begin')}}&nbsp;</el-button>
     </div>
 
+    <!-- scan pattern -->
+    <el-tag v-if="rule.pattern && rule.pattern.length" size="mini" style="margin-left: 10px;">
+      <i class="fa fa-search"></i> {{rule.pattern.join(' ')}}
+    </el-tag>
+
     <!-- key list -->
-    <ol class="del-batch-key-list-ol">
-      <li v-for="key, index in Object.keys(allKeys)" :key="index">{{ key }}</li>
-    </ol>
+    <RecycleScroller
+      class="del-batch-key-list"
+      :items="allKeysList"
+      :item-size="20"
+      key-field="str"
+      v-slot="{ item, index }"
+    >
+      <li>
+        <span class="list-index">{{ index + 1 }}.</span>
+        <span class="key-name" :title="item.str">{{ item.str }}</span>
+      </li>
+    </RecycleScroller>
   </el-card>
 </div>
 </template>
 
 <script type="text/javascript">
+import { RecycleScroller } from 'vue-virtual-scroller'
+
 export default {
   data() {
     return {
-      patternKeys: [],
       loadingScan: false,
       loadingDelete: false,
+      scanStreams: [],
+      allKeysList: [],
+      scanningEnd: false,
     };
   },
   props: ['client', 'rule', 'hotKeyScope'],
-  computed: {
-    allKeys() {
-      let dict = this.specifyKeys;
-
-      for (let key of this.patternKeys) {
-        dict[this.$util.bufToString(key)] = key;
-      }
-
-      return dict;
-    },
-    specifyKeys() {
-      let dict = {};
-
-      if (this.rule.key && this.rule.key.length) {
-        for (let key of this.rule.key) {
-          dict[this.$util.bufToString(key)] = key;
-        }
-      }
-
-      return dict;
-    },
-  },
+  components: { RecycleScroller },
   methods: {
     initKeys() {
-      this.patternKeys = [];
+      this.allKeysList = [];
+      this.rule.key && this.rule.key.length && this.addToList(this.rule.key);
 
       if (this.rule.pattern && this.rule.pattern.length) {
         this.loadingScan = true;
@@ -73,13 +74,20 @@ export default {
       nodes.map(node => {
         let scanOption = {
           match: pattern + '*',
-          count: 7000,
+          count: 20000,
         }
 
         let stream = node.scanBufferStream(scanOption);
+        this.scanStreams.push(stream);
 
         stream.on('data', keys => {
-          this.patternKeys = this.patternKeys.concat(keys.sort());
+          this.addToList(keys.sort());
+
+          // pause for dom rendering
+          stream.pause();
+          setTimeout(() => {
+            this.loadingScan && stream.resume();
+          }, 100);
         });
 
         stream.on('error', (e) => {
@@ -94,40 +102,108 @@ export default {
           // all nodes scan finished(cusor back to 0)
           if (--this.scanningCount <= 0) {
             this.loadingScan = false;
+            this.scanningEnd = true;
           }
         });
       });
     },
-    confirmDelete() {
-      this.loadingDelete = true;
+    addToList(keys) {
+      let list = [];
+      for (const key of keys) {
+        list.push({key: key, str: this.$util.bufToString(key)});
+      }
 
-      let keys = Object.values(this.allKeys);
-      let total = keys.length;
-      let last = total - 1;
+      this.allKeysList = this.allKeysList.concat(list);
+    },
+    toggleScanning(forcePause = null) {
+      this.loadingScan = (forcePause === null ? !this.loadingScan : !forcePause);
 
-      // one key per time instead of batch is for cluster...
-      for (let i = 0; i < total; i++) {
-        let promise = this.client.del(keys[i]);
-        // just wait the last one
-        if (i === last) {
-          promise.then((reply) => {
-            this.loadingDelete = false;
-
-            if (reply === 1) {
-              this.$message.success(this.$t('message.delete_success'));
-              this.$bus.$emit('removePreTab');
-              this.$bus.$emit('refreshKeyList', this.client);
-            }
-            else {
-              this.$message.error(this.$t('message.delete_failed'));
-            }
-          }).catch(e => {
-            this.loadingScan = false;
-            this.loadingDelete = false;
-            this.$message.error(e.message);
-          });
+      if (this.scanStreams.length) {
+        for (let stream of this.scanStreams) {
+          this.loadingScan ? stream.resume() : stream.pause();
         }
       }
+    },
+    confirmDelete() {
+      let keys = this.allKeysList;
+      let total = keys.length;
+
+      if (total <= 0) {
+        return;
+      }
+
+      this.loadingDelete = true;
+      let delPromise = null;
+
+      // standalone Redis, batch delete
+      if (!this.client.nodes) {
+        let chunked = [];
+        for (let i = 0; i < total; i++) {
+          chunked.push(keys[i].key);
+
+          // del 5000 keys one time
+          if (chunked.length >= 5000) {
+            delPromise = this.client.del(chunked);
+            chunked = [];
+          }
+        }
+
+        if (chunked.length) {
+          delPromise = this.client.del(chunked);
+        }
+        // use final promise
+        delPromise.then((reply) => {
+          if (reply > 0) {
+            this.afterDelete();
+          }
+          else {
+            this.deleteFailed(this.$t('message.delete_failed'));
+          }
+        }).catch(e => {
+          this.deleteFailed(e.message);
+        });
+      }
+
+      // cluster, one key per time instead of batch
+      else {
+        for (let i = 0; i < total; i++) {
+          delPromise = this.client.del(keys[i].key);
+          delPromise.catch(e => {});
+        }
+
+        // use final promise
+        delPromise.then((reply) => {
+          if (reply == 1) {
+            this.afterDelete();
+          }
+          else {
+            this.deleteFailed(this.$t('message.delete_failed'));
+          }
+        }).catch(e => {
+          this.deleteFailed(e.message);
+        });
+      }
+    },
+    afterDelete() {
+      this.loadingDelete = false;
+      this.allKeysList = [];
+
+      // empty the specified keys
+      // this.rule.key = [];
+
+      this.$message.success(this.$t('message.delete_success'));
+      this.$bus.$emit('refreshKeyList', this.client);
+
+      // except pattern mode scanning not to end, close pre tab
+      if (!this.rule.pattern.length || this.scanningEnd) {
+        this.$bus.$emit('removePreTab');
+      }
+    },
+    deleteFailed(msg = '') {
+      msg && this.$message.error(msg);
+
+      this.loadingScan = false;
+      this.loadingDelete = false;
     },
     initShortcut() {
       this.$shortcut.bind('ctrl+r, ⌘+r, f5', this.hotKeyScope, () => {
@@ -138,10 +214,13 @@ export default {
   },
   mounted() {
     this.initKeys();
-    this.initShortcut();
+    // disable f5 for streams on event cannot stop
+    // this.initShortcut();
   },
   beforeDestroy() {
-    this.$shortcut.deleteScope(this.hotKeyScope);
+    // this.$shortcut.deleteScope(this.hotKeyScope);
+    // cancel scanning
+    this.toggleScanning(true);
   },
 };
 </script>
@@ -153,28 +232,26 @@ export default {
     font-size: 120%;
   }
   .del-batch-card {
-    margin-top: 10px;
+    /*margin-top: 10px;*/
   }
-  .del-batch-key-list-ol {
-    min-height: calc(100vh - 272px);
+  .del-batch-key-list {
+    height: calc(100vh - 263px);
     overflow: auto;
     padding-left: 10px;
-    list-style-position: inside;
+    list-style: none;
+    margin-top: 10px;
   }
-  .del-batch-key-list-ol li {
+  .del-batch-key-list li {
     color: #333;
     font-size: 92%;
+    display: flex;
   }
-  .dark-mode .del-batch-key-list-ol li {
+  .dark-mode .del-batch-key-list li {
     color: #f7f7f7;
   }
-
-  .del-batch-card .el-loading-mask {
-    /*opacity: 0.3;*/
-    /*background: #263238;*/
-    background: rgba(38, 50, 56, .3);
-  }
-  .del-batch-card .el-loading-spinner {
-    top: 200px;
+  .del-batch-key-list li .key-name {
+    flex: 1;
+    overflow: hidden;
+    text-overflow: ellipsis;
   }
 </style>
