@@ -15,9 +15,11 @@
         <el-button type='text' icon="el-icon-edit-outline">{{$t('message.custom')}}</el-button>
       </el-option>
     </el-select>
-    <span @click='copyContent' :title='$t("message.copy")' class='el-icon-document formater-copy-icon'>{{$t("message.copy")}}</span>
-    <span v-if='!contentVisible' class='formater-binary-tag'>[Hex]</span>
-    <span class='formater-binary-tag'>Size: {{ $util.humanFileSize(buffSize) }}</span>
+    <el-tag v-if='!contentVisible' size="mini" class='formater-binary-tag' :disable-transitions='true'>[Hex]</el-tag>
+    <el-tag class='formater-binary-tag' size="mini" :disable-transitions='true'>Size: {{ $util.humanFileSize(buffSize) }}</el-tag>
+    <el-button  @click='copyContent' :title='$t("message.copy")' type='text' size='mini'>
+      <i class="el-icon-document"></i>{{$t("message.copy")}}
+    </el-button>
     <br>
 
     <component
@@ -26,7 +28,6 @@
       :content='content'
       :name="selectedView"
       :contentVisible='contentVisible'
-      :textrows='textrows'
       :disabled='disabled'
       :redisKey="redisKey"
       :dataMap="dataMap">
@@ -42,9 +43,13 @@ import ViewerJson from '@/components/ViewerJson';
 import ViewerBinary from '@/components/ViewerBinary';
 import ViewerUnserialize from '@/components/ViewerUnserialize';
 import ViewerBrotli from '@/components/ViewerBrotli';
+import ViewerGzip from '@/components/ViewerGzip';
+import ViewerDeflate from '@/components/ViewerDeflate';
 import ViewerMsgpack from '@/components/ViewerMsgpack';
 import ViewerOverSize from '@/components/ViewerOverSize';
 import ViewerCustom from '@/components/ViewerCustom';
+import ViewerProtobuf from '@/components/ViewerProtobuf';
+import ViewerDeflateRaw from '@/components/ViewerDeflateRaw';
 
 export default {
   data() {
@@ -59,21 +64,39 @@ export default {
         { value: 'ViewerMsgpack', text: 'Msgpack' },
         { value: 'ViewerUnserialize', text: 'Unserialize' },
         { value: 'ViewerBrotli', text: 'Brotli' },
+        { value: 'ViewerGzip', text: 'Gzip' },
+        { value: 'ViewerDeflate', text: 'Deflate' },
+        { value: 'ViewerDeflateRaw', text: 'DeflateRaw' },
+        { value: 'ViewerProtobuf', text: 'Protobuf' },
       ],
       selectStyle: {
         float: this.float,
       },
       overSizeBytes: 20971520, // 20MB
+      autoFormated: false,
     };
   },
-  components: {ViewerText, ViewerHex, ViewerJson, ViewerBinary, ViewerUnserialize, ViewerMsgpack, ViewerOverSize, ViewerCustom, ViewerBrotli},
+  components: {
+    ViewerText,
+    ViewerHex,
+    ViewerJson,
+    ViewerBinary,
+    ViewerUnserialize,
+    ViewerMsgpack,
+    ViewerOverSize,
+    ViewerCustom,
+    ViewerBrotli,
+    ViewerGzip,
+    ViewerDeflate,
+    ViewerProtobuf,
+    ViewerDeflateRaw,
+  },
   props: {
-    float: {default: 'right'},
-    content: {default: () => Buffer.from('')},
-    textrows: {default: 6},
-    disabled: {type: Boolean, default: false},
-    redisKey:  {default: () => Buffer.from('')},
-    dataMap: {type: Object, default: () => {}},
+    float: { default: 'right' },
+    content: { default: () => Buffer.from('') },
+    disabled: { type: Boolean, default: false },
+    redisKey: { default: () => Buffer.from('') },
+    dataMap: { type: Object, default: () => {} },
   },
   computed: {
     contentVisible() {
@@ -91,9 +114,9 @@ export default {
     },
     viewersMap() {
       // add oversize tmp
-      let map = {OverSize: 'ViewerOverSize'};
+      const map = { OverSize: 'ViewerOverSize' };
 
-      this.viewers.forEach(item => {
+      this.viewers.forEach((item) => {
         map[item.text] = item.value;
       });
 
@@ -108,7 +131,13 @@ export default {
   },
   watch: {
     content() {
+      // auto format only when first in #920
+      if (this.autoFormated) {
+        return;
+      }
+
       this.autoFormat();
+      this.autoFormated = true;
     },
     selectedView(viewer) {
       // custom viewer com may same, force change
@@ -148,29 +177,45 @@ export default {
         return this.changeViewer('Json');
       }
       // php unserialize
-      else if (this.$util.isPHPSerialize(this.content)) {
+      if (this.$util.isPHPSerialize(this.content)) {
         return this.changeViewer('Unserialize');
       }
       // msgpack
-      else if (this.$util.isMsgpack(this.content)) {
+      if (this.$util.isMsgpack(this.content)) {
         return this.changeViewer('Msgpack');
       }
       // brotli unserialize
-      else if (this.$util.isBrotli(this.content)) {
+      if (this.$util.isBrotli(this.content)) {
         return this.changeViewer('Brotli');
       }
+      // gzip
+      if (this.$util.isGzip(this.content)) {
+        return this.changeViewer('Gzip');
+      }
+      // deflate
+      if (this.$util.isDeflate(this.content)) {
+        return this.changeViewer('Deflate');
+      }
+      // protobuf
+      if (this.$util.isProtobuf(this.content)) {
+        return this.changeViewer('Protobuf');
+      }
+      // deflateRaw
+      if (this.$util.isDeflateRaw(this.content)) {
+        return this.changeViewer('DeflateRaw');
+      }
+
       // hex
-      else if (!this.contentVisible) {
+      if (!this.contentVisible) {
         return this.changeViewer('Hex');
       }
-      else {
-        return this.changeViewer('Text');
-      }
+
+      return this.changeViewer('Text');
     },
     copyContent() {
-      let content = (typeof this.$refs.viewer.copyContent == 'function') ?
-                    this.$refs.viewer.copyContent() :
-                    this.content;
+      const content = (typeof this.$refs.viewer.copyContent === 'function')
+        ? this.$refs.viewer.copyContent()
+        : this.content;
 
       this.$util.copyToClipboard(content);
       this.$message.success(this.$t('message.copy_success'));
@@ -182,14 +227,12 @@ export default {
         return;
       }
 
-      formatters.forEach(formatter => {
-        this.viewers.push({value: 'ViewerCustom', text: formatter.name, type: 'custom'});
+      formatters.forEach((formatter) => {
+        this.viewers.push({ value: 'ViewerCustom', text: formatter.name, type: 'custom' });
       });
     },
     removeCustom() {
-      this.viewers = this.viewers.filter(item => {
-        return item.type !== 'custom';
-      });
+      this.viewers = this.viewers.filter(item => item.type !== 'custom');
     },
   },
   mounted() {
@@ -201,21 +244,25 @@ export default {
 
 <style type="text/css">
   .format-selector {
-    width: 122px;
+    width: 130px;
   }
   .format-selector .el-input__inner {
     height: 22px !important;
   }
 
+  /*outline same with text viewer's .el-textarea__inner*/
   .text-formated-container {
     border: 1px solid #dcdfe6;
-    min-height: 114px;
-    padding: 5px 15px;
-    line-height: 1.5;
-    border-radius: 5px;
+    padding: 5px 10px;
+    border-radius: 4px;
   }
   .dark-mode .text-formated-container {
     border-color: #7f8ea5;
+  }
+
+  .format-viewer-container textarea {
+    min-height: 194px !important;
+    height: calc(100vh - 686px);
   }
 
   .collapse-container {
@@ -227,13 +274,6 @@ export default {
     padding: 9px 0;
   }
   .formater-binary-tag {
-    /*padding-left: 5px;*/
-    color: #7ab3ef;
-    font-size: 80%;
-  }
-  .formater-copy-icon {
-    color: #7ab3ef;
-    cursor: pointer;
     font-size: 80%;
   }
 </style>
