@@ -2,9 +2,9 @@
   <div>
     <div>
       <!-- add button -->
-      <el-form :inline="true" size="small">
+      <el-form :inline="true">
         <el-form-item>
-          <el-button size="small" type="primary" @click='showEditDialog({})'>{{ $t('message.add_new_line') }}</el-button>
+          <el-button type="primary" @click='showEditDialog({})'>{{ $t('message.add_new_line') }}</el-button>
         </el-form-item>
       </el-form>
 
@@ -26,8 +26,8 @@
     <!-- content table -->
     <el-table
       stripe
-      size="small"
       border
+      size='mini'
       min-height=300
       :data="listData">
       <el-table-column
@@ -48,10 +48,19 @@
       </el-table-column>
 
       <el-table-column label="Operation">
+        <template slot="header" slot-scope="scope">
+          <input
+            class="el-input__inner key-detail-filter-value"
+            v-model="filterValue"
+            @keyup.enter='initShow()'
+            :placeholder="$t('message.key_to_search')"/>
+          <i :class='loadingIcon'></i>
+        </template>
         <template slot-scope="scope">
           <el-button type="text" @click="$util.copyToClipboard(scope.row.value)" icon="el-icon-document" :title="$t('message.copy')"></el-button>
           <el-button type="text" @click="showEditDialog(scope.row)" icon="el-icon-edit" :title="$t('message.edit_line')"></el-button>
           <el-button type="text" @click="deleteLine(scope.row)" icon="el-icon-delete" :title="$t('el.upload.delete')"></el-button>
+          <el-button type="text" @click="dumpCommand(scope.row)" icon="fa fa-code" :title="$t('message.dump_to_clipboard')"></el-button>
         </template>
       </el-table-column>
     </el-table>
@@ -87,8 +96,9 @@ export default {
       beforeEditItem: {},
       editLineItem: {},
       loadingIcon: '',
-      pageSize: 30,
+      pageSize: 100,
       pageIndex: 0,
+      oneTimeListLength: 0,
       loadMoreDisable: false,
     };
   },
@@ -105,30 +115,59 @@ export default {
       resetTable && this.resetTable();
       this.loadingIcon = 'el-icon-loading';
 
-      let start = this.pageSize * this.pageIndex;
-      let end = start + this.pageSize - 1;
+      // scan
+      this.listScan();
+      // total lines
+      this.initTotal();
+    },
+    listScan() {
+      const filterValue = this.filterValue;
+      const pageSize = filterValue ? 500 : this.pageSize;
+
+      let start = pageSize * this.pageIndex;
+      let end = start + pageSize - 1;
 
       this.client.lrangeBuffer([this.redisKey, start, end]).then((reply) => {
-        let listData = [];
+        // scanning end
+        if (!reply || !reply.length) {
+          this.loadingIcon = '';
+          this.loadMoreDisable = true;
+          return;
+        }
 
+        const listData = [];
         for (const i of reply) {
+          if (filterValue) {
+            if (!i.includes(filterValue)) {
+              continue;
+            }
+          }
+
           listData.push({
             value: i,
-            // valueDisplay: this.$util.bufToString(i),
+            uniq: Math.random(),
           });
         }
 
-        this.listData = resetTable ? listData : this.listData.concat(listData);
-        (listData.length < this.pageSize) && (this.loadMoreDisable = true);
-        this.loadingIcon = '';
+        this.oneTimeListLength += listData.length;
+        this.listData = this.listData.concat(listData);
+
+        if (this.oneTimeListLength >= this.pageSize) {
+          this.loadingIcon = '';
+          this.oneTimeListLength = 0;
+          return;
+        }
+
+        if (this.cancelScanning) {
+          return;
+        }
+        // continue scanning until to pagesize
+        this.loadMore();
       }).catch(e => {
         this.loadingIcon = '';
         this.loadMoreDisable = true;
         this.$message.error(e.message);
       });
-
-      // total lines
-      this.initTotal();
     },
     initTotal() {
       this.client.llen(this.redisKey).then((reply) => {
@@ -138,21 +177,34 @@ export default {
     resetTable() {
       this.listData = [];
       this.pageIndex = 0;
+      this.oneTimeListLength = 0;
       this.loadMoreDisable = false;
     },
     loadMore() {
       this.pageIndex++;
-      this.initShow(false);
+      this.listScan();
     },
     openDialog() {
-      // this.$nextTick(() => {
-      //   this.$refs.formatViewer.autoFormat();
-      // });
+      this.$nextTick(() => {
+        this.$refs.formatViewer.autoFormat();
+      });
     },
     showEditDialog(row) {
       this.editLineItem = row;
       this.beforeEditItem = this.$util.cloneObjWithBuff(row);
       this.editDialog = true;
+
+      this.rowUniq = row.uniq;
+    },
+    dumpCommand(item) {
+      const lines = item ? [item] : this.listData;
+      const params = lines.map(line => {
+        return this.$util.bufToQuotation(line.value);
+      });
+
+      const command = `RPUSH ${this.$util.bufToQuotation(this.redisKey)} ${params.join(' ')}`;
+      this.$util.copyToClipboard(command);
+      this.$message.success({message: this.$t('message.copy_success'), duration: 800});
     },
     editLine() {
       const key = this.redisKey;
@@ -179,17 +231,19 @@ export default {
         if (reply > 0) {
           // edit key remove previous value
           if (before.value) {
-            client.lrem(
-              key,
-              1,
-              before.value
-            ).then((reply) => {
-              this.initShow();
-            });
+            client.lrem(key, 1, before.value);
           }
 
+          // this.initShow(); // do not reinit, #786
+          const newLine = {value: afterValue, uniq: Math.random()};
+          // edit line
+          if (this.rowUniq) {
+            this.$util.listSplice(this.listData, this.rowUniq, newLine);
+          }
+          // new line
           else {
-            this.initShow();
+            this.listData.push(newLine);
+            this.total++;
           }
 
           this.$message.success({
@@ -209,13 +263,15 @@ export default {
           1,
           row.value
         ).then((reply) => {
-          if (reply === 1) {
+          if (reply == 1) {
             this.$message.success({
               message: this.$t('message.delete_success'),
               duration: 1000,
             });
 
-            this.initShow();
+            // this.initShow(); // do not reinit, #786
+            this.$util.listSplice(this.listData, row.uniq);
+            this.total--;
           }
         }).catch(e => {this.$message.error(e.message);});
       }).catch(() => {});
@@ -223,6 +279,9 @@ export default {
   },
   mounted() {
     this.initShow();
+  },
+  beforeDestroy() {
+    this.cancelScanning = true;
   },
 };
 </script>
