@@ -7,8 +7,7 @@
       :content='content'
       :binary='binary'
       :redisKey='redisKey'
-      float=''
-      :textrows=12>
+      float=''>
     </FormatViewer>
   </el-form-item>
 
@@ -92,6 +91,12 @@ export default {
         return false;
       });
     },
+    dumpCommand() {
+      const command = `SET ${this.$util.bufToQuotation(this.redisKey)} ` +
+                      this.$util.bufToQuotation(this.content);
+      this.$util.copyToClipboard(command);
+      this.$message.success({message: this.$t('message.copy_success'), duration: 800});
+    },
   },
   mounted() {
     this.initShow();
@@ -111,12 +116,7 @@ export default {
     height: calc(100vh - 286px);
   }
   /*json in monaco editor*/
-  .key-content-string #monaco-editor-con {
-    height: calc(100vh - 331px);
-  }
-  /*not text viewer box, such as json*/
-  .key-content-string .text-formated-container {
-    box-sizing: border-box;
-    min-height: calc(100vh - 286px);
+  .key-content-string .text-formated-container .monaco-editor-con {
+    height: calc(100vh - 330px);
   }
 </style>
