@@ -2,9 +2,9 @@
   <div>
     <div>
       <!-- add button -->
-      <el-form :inline="true" size="small">
+      <el-form :inline="true">
         <el-form-item>
-          <el-button size="small" type="primary" @click="showEditDialog({})">{{ $t('message.add_new_line') }}</el-button>
+          <el-button type="primary" @click="showEditDialog({})">{{ $t('message.add_new_line') }}</el-button>
         </el-form-item>
       </el-form>
 
@@ -29,8 +29,8 @@
     <!-- content table -->
     <el-table
       stripe
-      size="small"
       border
+      size='mini'
       min-height=300
       :data="zsetData">
       <el-table-column
@@ -70,6 +70,7 @@
           <el-button type="text" @click="$util.copyToClipboard(scope.row.member)" icon="el-icon-document" :title="$t('message.copy')"></el-button>
           <el-button type="text" @click="showEditDialog(scope.row)" icon="el-icon-edit" :title="$t('message.edit_line')"></el-button>
           <el-button type="text" @click="deleteLine(scope.row)" icon="el-icon-delete" :title="$t('el.upload.delete')"></el-button>
+          <el-button type="text" @click="dumpCommand(scope.row)" icon="fa fa-code" :title="$t('message.dump_to_clipboard')"></el-button>
         </template>
       </el-table-column>
     </el-table>
@@ -105,7 +106,7 @@ export default {
       beforeEditItem: {},
       editLineItem: {},
       loadingIcon: '',
-      pageSize: 30,
+      pageSize: 100,
       pageIndex: 0,
       searchPageSize: 1000,
       oneTimeListLength: 0,
@@ -133,7 +134,7 @@ export default {
 
       // default mode, ordered
       else {
-        this.getListRange();
+        this.getListRange(resetTable);
         this.pageIndex++;
       }
 
@@ -146,6 +147,8 @@ export default {
       }).catch(e => {});
     },
     resetTable() {
+      // stop scanning first, #815
+      this.scanStream && this.scanStream.pause();
       this.zsetData = [];
       this.pageIndex = 0;
       this.scanStream = null;
@@ -222,6 +225,7 @@ export default {
           score: Number(list[i + 1]),
           member: list[i],
           // memberDisplay: this.$util.bufToString(list[i]),
+          uniq: Math.random(),
         });
       }
 
@@ -231,14 +235,27 @@ export default {
       return this.filterValue ? `*${this.filterValue}*` : '*';
     },
     openDialog() {
-      // this.$nextTick(() => {
-      //   this.$refs.formatViewer.autoFormat();
-      // });
+      this.$nextTick(() => {
+        this.$refs.formatViewer.autoFormat();
+      });
     },
     showEditDialog(row) {
       this.editLineItem = row;
       this.beforeEditItem = this.$util.cloneObjWithBuff(row);
       this.editDialog = true;
+
+      this.rowUniq = row.uniq;
+    },
+    dumpCommand(item) {
+      const lines = item ? [item] : this.zsetData;
+      const params = lines.map(line => {
+        return `${String(line.score)} ` +
+                this.$util.bufToQuotation(line.member);
+      });
+
+      const command = `ZADD ${this.$util.bufToQuotation(this.redisKey)} ${params.join(' ')}`;
+      this.$util.copyToClipboard(command);
+      this.$message.success({message: this.$t('message.copy_success'), duration: 800});
     },
     editLine() {
       const key = this.redisKey;
@@ -261,20 +278,23 @@ export default {
       ).then((reply) => {
         // edit key member changed
         if (before.member && !before.member.equals(afterMember)) {
-          client.zrem(
-            key,
-            before.member
-          ).then((reply) => {
-            this.initShow();
-          });
+          client.zrem(key, before.member);
         }
 
+        // this.initShow(); // do not reinit, #786
+        const newLine = {score: afterScore, member: afterMember, uniq: Math.random()};
+        // edit line
+        if (this.rowUniq) {
+          this.$util.listSplice(this.zsetData, this.rowUniq, newLine);
+        }
+        // new line
         else {
-          this.initShow();
+          this.zsetData.push(newLine);
+          this.total++;
         }
 
         this.$message.success({
-          message: reply ? this.$t('message.add_success') : this.$t('message.modify_success'),
+          message: reply == 1 ? this.$t('message.add_success') : this.$t('message.modify_success'),
           duration: 1000,
         });
       }).catch(e => {this.$message.error(e.message);});
@@ -288,13 +308,15 @@ export default {
           this.redisKey,
           row.member
         ).then((reply) => {
-          if (reply === 1) {
+          if (reply == 1) {
             this.$message.success({
               message: this.$t('message.delete_success'),
               duration: 1000,
             });
 
-            this.initShow();
+            // this.initShow(); // do not reinit, #786
+            this.$util.listSplice(this.zsetData, row.uniq);
+            this.total--;
           }
         }).catch(e => {this.$message.error(e.message);});
       }).catch(() => {});
