@@ -8,6 +8,7 @@
         :redisKey="redisKey"
         :keyType="keyType"
         @refreshContent='refreshContent'
+        @dumpCommand='dumpCommand'
         :hotKeyScope='hotKeyScope'
         class="key-header-info">
       </KeyHeader>
@@ -58,7 +59,6 @@ export default {
         set   : 'KeyContentSet',
         list  : 'KeyContentList',
         stream  : 'KeyContentStream',
-        stream  : 'KeyContentStream',
         'ReJSON-RL': 'KeyContentReJson',
       };
 
@@ -73,7 +73,9 @@ export default {
     },
     refreshContent() {
       this.client.exists(this.redisKey).then(reply => {
-        if (!reply) {
+        if (reply == 0) {
+          // clear interval if auto refresh opened
+          this.$refs.keyHeader.removeInterval();
           return this.$message.error(this.$t('message.key_not_exists'));
         }
 
@@ -82,13 +84,16 @@ export default {
         this.$message.error('Exists Error: ' + e.message);
       });
     },
+    dumpCommand() {
+      this.$refs.keyContent && this.$refs.keyContent.dumpCommand();
+    },
   },
 };
 </script>
 
 <style type="text/css">
   .key-tab-container {
-    padding-left: 5px;
+    /*padding-left: 5px;*/
   }
   .key-header-info {
     margin-top: 15px;
@@ -113,5 +118,21 @@ export default {
   }
   .content-more-container .content-more-btn {
     width: 95%;
+    padding-top: 5px;
+    padding-bottom: 5px;
+  }
+
+  /*data table list styles*/
+  .key-content-container .el-table {
+    border-radius: 3px;
+  }
+  /*table list height*/
+  .key-content-container .el-table .el-table__body td {
+    padding: 0px 0px;
+  }
+
+  /*table list border*/
+  .key-content-container .el-table--border td, .key-content-container .el-table--border th {
+    border-right-width: 0;
   }
 </style>
