@@ -1,52 +1,74 @@
 <template>
   <div>
-    <el-form :inline="true">
-      <!-- key name -->
-      <el-col :span=12 class='key-header-item key-name-input'>
-        <el-input
-          ref="keyNameInput"
-          :value="$util.bufToString(keyName)"
-          @change='changeKeyInput'
-          @keyup.enter.native="renameKey"
+    <!-- key name -->
+    <div class="key-header-item key-name-input">
+      <el-input
+        ref="keyNameInput"
+        :value="$util.bufToString(keyName)"
+        @change='changeKeyInput'
+        @keyup.enter.native="renameKey"
+        :title="$t('message.click_enter_to_rename')"
+        placeholder="KeyName">
+        <span slot="prepend" class="key-detail-type">{{ keyType }}</span>
+        <i class="el-icon-check el-input__icon cursor-pointer"
+          slot="suffix"
           :title="$t('message.click_enter_to_rename')"
-          placeholder="KeyName">
-          <span slot="prepend" class="key-detail-type">{{ keyType }}</span>
-          <i class="el-icon-check el-input__icon cursor-pointer"
-            slot="suffix"
-            :title="$t('message.click_enter_to_rename')"
-            @click="renameKey">
-          </i>
-        </el-input>
-      </el-col>
+          @click="renameKey">
+        </i>
+      </el-input>
+    </div>
 
-      <!-- key ttl -->
-      <el-col :span=6 class='key-header-item key-ttl-input'>
-        <el-input
-          v-model="keyTTL"
-          @keyup.enter.native="ttlKey"
-          :title="$t('message.click_enter_to_ttl')">
-          <span slot="prepend">TTL</span>
-          <!-- remove expire -->
-          <i class="el-icon-close el-input__icon cursor-pointer"
-            slot="suffix"
-            :title="$t('message.persist')"
-            @click="persistKet">
-          </i>
-          <!-- save ttl -->
-          <i class="el-icon-check el-input__icon cursor-pointer"
-            slot="suffix"
-            :title="$t('message.click_enter_to_ttl')"
-            @click="ttlKey">
-          </i>
-        </el-input>
-      </el-col>
+    <!-- key ttl -->
+    <div class="key-header-item key-ttl-input">
+      <el-input
+        type="number"
+        v-model="keyTTL"
+        @keyup.enter.native="ttlKey"
+        :title="$util.leftTime(keyTTL)">
+        <span slot="prepend">TTL</span>
+        <!-- remove expire -->
+        <i class="el-icon-close el-input__icon cursor-pointer"
+          slot="suffix"
+          :title="$t('message.persist')"
+          @click="persistKet">
+        </i>
+        <!-- save ttl -->
+        <i class="el-icon-check el-input__icon cursor-pointer"
+          slot="suffix"
+          :title="$t('message.click_enter_to_ttl')"
+          @click="ttlKey">
+        </i>
+      </el-input>
+    </div>
 
-      <!-- del & refresh btn -->
-      <el-col :span=6 class='key-header-item key-header-btn-con'>
-        <el-button ref='deleteBtn' type="danger" @click="deleteKey" icon="el-icon-delete" :title="$t('el.upload.delete')+' Ctrl+d'"></el-button>
-        <el-button ref='refreshBtn' type="success" @click="refreshKey" icon="el-icon-refresh" :title="$t('message.refresh_connection')+' Ctrl+r / F5'"></el-button>
-      </el-col>
-    </el-form>
+    <!-- del & refresh btn -->
+    <div class='key-header-item key-header-btn-con'>
+      <!-- del btn -->
+      <el-button ref='deleteBtn' type="danger" @click="deleteKey" icon="el-icon-delete" :title="$t('el.upload.delete')+' Ctrl+d'"></el-button>
+      <!-- refresh btn -->
+      <!-- <el-button ref='refreshBtn' type="success" @click="refreshKey" icon="el-icon-refresh" :title="$t('message.refresh_connection')+' Ctrl+r / F5'"></el-button> -->
+
+      <!-- refresh btn component -->
+      <el-popover
+        placement="bottom"
+        :open-delay="500"
+        trigger="hover">
+        <el-tag type="info">
+          <i class="el-icon-refresh"></i>
+          {{ $t('message.auto_refresh') }}
+        </el-tag>
+
+        <el-tooltip :content="$t('message.auto_refresh_tip', {interval: refreshInterval / 1000})" effect="dark" placement="bottom">
+          <el-switch v-model='autoRefresh' @change="refreshInit"></el-switch>
+        </el-tooltip>
+
+        <!-- refresh btn -->
+        <el-button slot="reference" ref='refreshBtn' type="success" @click="refreshKey" icon="el-icon-refresh" :title="$t('message.refresh_connection')+' Ctrl+r / F5'" :class="autoRefresh?'rotating':''"></el-button>
+      </el-popover>
+
+      <!-- dump btn -->
+      <el-button ref='dumpBtn' type="primary" @click="dumpCommand" icon="fa fa-code" :title="$t('message.dump_to_clipboard')"></el-button>
+    </div>  
   </div>
 </template>
 
@@ -57,6 +79,8 @@ export default {
       keyName: this.redisKey,
       keyTTL: -1,
       binary: false,
+      autoRefresh: false,
+      refreshInterval: 2000,
     };
   },
   props: ['client', 'redisKey', 'keyType', 'hotKeyScope'],
@@ -82,6 +106,24 @@ export default {
       this.initShow();
       this.$emit('refreshContent');
     },
+    refreshInit() {
+      this.refreshTimer && clearInterval(this.refreshTimer);
+
+      if (this.autoRefresh) {
+        this.refreshKey();
+
+        this.refreshTimer = setInterval(() => {
+          this.refreshKey();
+        }, this.refreshInterval);
+      }
+    },
+    removeInterval() {
+      this.autoRefresh = false;
+      this.refreshInit();
+    },
+    dumpCommand() {
+      this.$emit('dumpCommand');
+    },
     deleteKey() {
       this.$confirm(
         this.$t('message.confirm_to_delete_key', { key: this.$util.bufToString(this.redisKey) }),
@@ -89,7 +131,7 @@ export default {
       )
       .then(() => {
         this.client.del(this.redisKey).then((reply) => {
-          if (reply === 1) {
+          if (reply == 1) {
             this.$message.success({
               message: this.$t('message.delete_success'),
               duration: 1000,
@@ -156,7 +198,7 @@ export default {
     },
     setTTL(keyDeleted = false) {
       this.client.expire(this.redisKey, this.keyTTL).then((reply) => {
-        if (reply) {
+        if (reply == 1) {
           this.$message.success({
             message: this.$t('message.modify_success'),
             duration: 1000,
@@ -203,6 +245,7 @@ export default {
     this.initShortcut();
   },
   beforeDestroy() {
+    clearInterval(this.refreshTimer);
     this.$shortcut.deleteScope(this.hotKeyScope);
   },
 };
@@ -220,19 +263,36 @@ export default {
   }
 
   .key-header-item {
-    padding-right: 15px;
-    margin-bottom: 10px;
+    /*padding-right: 15px;*/
+    /*margin-bottom: 10px;*/
+    float: left;
   }
 
   .key-header-item.key-name-input {
-    min-width: 317px;
-    max-width: 650;
+    width: calc(100% - 402px);
+    min-width: 220px;
+    max-width: 800px;
+    margin-right: 15px;
+    margin-bottom: 10px;
   }
   .key-header-item.key-ttl-input {
-    min-width: 200px;
-    max-width: 400px;
+    width: 220px; 
+    margin-right: 15px; 
+    margin-bottom: 10px;
   }
-  .key-header-item.key-header-btn-con {
-    width: 130px;
+  /*hide number input button*/
+  .key-header-item.key-ttl-input input::-webkit-inner-spin-button,
+  .key-header-item.key-ttl-input input::-webkit-outer-spin-button {
+    appearance: none;
   }
+
+  .key-header-item.key-header-btn-con .el-button+.el-button {
+    margin-left: 4px;
+  }
+
+  /*refresh btn rotating*/
+  .key-header-info .key-header-btn-con .rotating .el-icon-refresh{
+    animation: rotate 1.5s linear infinite;
+  }
+
 </style>
