@@ -10,7 +10,7 @@
           </el-form-item>
 
           <el-form-item :label="$t('message.password')">
-            <el-input v-model="connection.auth" type='password' autocomplete="off"></el-input>
+            <InputPassword v-model="connection.auth" placeholder="Auth"></InputPassword>
           </el-form-item>
 
           <el-form-item :label="$t('message.connection_name')">
@@ -55,6 +55,13 @@
             {{ $t('message.cluster_faq') }}
           </el-popover>
         </el-checkbox>
+        <el-checkbox v-model="connection.connectionReadOnly">
+          Readonly
+          <el-popover trigger="hover">
+            <i slot="reference" class="el-icon-question"></i>
+            {{ $t('message.connection_readonly') }}
+          </el-popover>
+        </el-checkbox>
       </el-form-item>
     </el-form>
 
@@ -76,18 +83,15 @@
           </el-form-item>
 
           <el-form-item :label="$t('message.private_key')">
-            <el-tooltip effect="dark">
-              <div slot="content" v-html="$t('message.private_key_faq')"></div>
-              <FileInput
-                :file.sync='connection.sshOptions.privatekey'
-                :bookmark.sync='connection.sshOptions.privatekeybookmark'
-                placeholder='SSH Private Key'>
-              </FileInput>
-            </el-tooltip>
+            <FileInput
+              :file.sync='connection.sshOptions.privatekey'
+              :bookmark.sync='connection.sshOptions.privatekeybookmark'
+              placeholder='SSH Private Key'>
+            </FileInput>
           </el-form-item>
 
           <el-form-item label="Passphrase">
-            <el-input v-model="connection.sshOptions.passphrase" type='password' autocomplete="off" placeholder='Passphrase for Private Key'></el-input>
+            <InputPassword v-model="connection.sshOptions.passphrase" placeholder='Passphrase for Private Key'></InputPassword>
           </el-form-item>
         </el-col>
 
@@ -98,7 +102,7 @@
           </el-form-item>
 
           <el-form-item :label="$t('message.password')">
-            <el-input v-model="connection.sshOptions.password" type='password' autocomplete="off"></el-input>
+            <InputPassword v-model="connection.sshOptions.password" placeholder="SSH Password"></InputPassword>
           </el-form-item>
 
           <el-form-item :label="$t('message.timeout')">
@@ -157,7 +161,7 @@
         <!-- left col -->
         <el-col :span=12>
           <el-form-item :label="$t('message.redis_node_password')">
-            <el-input type='password' v-model="connection.sentinelOptions.nodePassword" autocomplete="off" placeholder='Redis Node Password'></el-input>
+            <InputPassword v-model="connection.sentinelOptions.nodePassword" placeholder='Redis Node Password'></InputPassword>
           </el-form-item>
         </el-col>
 
@@ -180,6 +184,7 @@
 <script type="text/javascript">
 import storage from '@/storage';
 import FileInput from '@/components/FileInput';
+import InputPassword from '@/components/InputPassword';
 
 export default {
   data() {
@@ -195,6 +200,7 @@ export default {
         name: '',
         separator: ':',
         cluster: false,
+        connectionReadOnly: false,
         sshOptions: {
           host: '',
           port: 22,
@@ -220,7 +226,7 @@ export default {
       sentinelOptionsShow: false,
     }
   },
-  components: {FileInput},
+  components: {FileInput, InputPassword},
   props: {
     config: {
       default: _ => new Array,
