@@ -1,139 +1,168 @@
 <template>
   <!-- setting dialog -->
-  <el-dialog :title="$t('message.settings')" :visible.sync="visible">
+  <el-dialog :title="$t('message.settings')" :visible.sync="visible" custom-class="setting-main-dialog">
     <el-form label-position="top" size="mini">
 
-      <!-- theme select-->
-      <el-form-item :label="$t('message.dark_mode')">
-        <el-switch v-model='darkMode' @change="changeTheme"></el-switch>
-      </el-form-item>
+      <el-card :header="$t('message.ui_settings')" class="setting-card">
+        <el-row :gutter="10" justify="space-between" type="flex" class="setting-row">
+          <el-col :sm="12" :lg="5">
+            <!-- theme select-->
+            <el-form-item :label="$t('message.dark_mode')">
+              <el-switch v-model='darkMode' @change="changeTheme"></el-switch>
+            </el-form-item>
+          </el-col>
+          <el-col :sm="12" :lg="7">
+            <!-- language select -->
+            <el-form-item :label="$t('message.select_lang')">
+              <LanguageSelector></LanguageSelector>
+            </el-form-item>
+          </el-col>
+          <el-col :sm="12" :lg="5">
+            <!-- zoom page -->
+            <el-form-item :label="$t('message.page_zoom')">
+              <el-input-number
+                size="mini"
+                placeholder='1.0'
+                :min=0.5
+                :max=2.0
+                :step=0.1
+                :precision=1
+                @change='changeZoom'
+                v-model='form.zoomFactor'>
+              </el-input-number>
+            </el-form-item>
+          </el-col>
+          <el-col :sm="12" :lg="7">
+            <!-- font-family -->
+            <el-form-item :label="$t('message.font_family')">
+              <span slot="label">
+                {{ $t('message.font_family') }}
+                <el-popover
+                  placement="top-start"
+                  :title="$t('message.font_faq_title')"
+                  trigger="hover">
+                  <i slot="reference" class="el-icon-question"></i>
+                  <p v-html="$t('message.font_faq')"></p>
+                </el-popover>
+                <i v-if="loadingFonts" class="el-icon-loading"></i>
+              </span>
+              <!-- font-family select -->
+              <el-select v-model="form.fontFamily" @visible-change="getAllFonts" allow-create default-first-option
+                         filterable multiple class="setting-font-select">
+                <el-option
+                  v-for="(font, index) in allFonts"
+                  :key="index"
+                  :label="font"
+                  :value="font">
+                  <!-- for better performance, do not display font-family -->
+                  <!-- :style="{'font-family': font}"> -->
+                </el-option>
+              </el-select>
+            </el-form-item>
+          </el-col>
+        </el-row>
+      </el-card>
 
-      <!-- language select -->
-      <el-form-item :label="$t('message.select_lang')">
-        <LanguageSelector></LanguageSelector>
-      </el-form-item>
+      <el-card :header="$t('message.common_settings')" class="setting-card">
+        <el-row :gutter="20" justify="space-between" type="flex" class="setting-row">
+          <el-col :sm="12" :lg="12">
+            <!-- keys per loading -->
+            <el-form-item>
+              <el-input-number
+                size="mini"
+                placeholder='500'
+                :min=10
+                :max=20000
+                :step=50
+                v-model='form.keysPageSize'>
+              </el-input-number>&nbsp;
+              <!-- load all switch -->
+              <!-- <el-switch v-model='form.showLoadAllKeys'></el-switch>
+              {{ $t('message.show_load_all_keys') }} -->
 
-      <!-- zoom page -->
-      <el-form-item :label="$t('message.page_zoom')">
-        <el-input-number
-          size="mini"
-          placeholder='1.0'
-          :min=0.5
-          :max=2.0
-          :step=0.1
-          :precision=1
-          @change='changeZoom'
-          v-model='form.zoomFactor'>
-        </el-input-number>
-      </el-form-item>
+              <span slot="label">
+                {{ $t('message.keys_per_loading') }}
+                <el-popover
+                  :content="$t('message.keys_per_loading_tip')"
+                  placement="top-start"
+                  trigger="hover">
+                  <i slot="reference" class="el-icon-question"></i>
+                </el-popover>
+              </span>
+            </el-form-item>
+          </el-col>
+          <el-col :sm="12" :lg="12">
+            <!-- export connections -->
+            <el-form-item :label="$t('message.config_connections')">
+              <el-button icon="el-icon-upload2" @click="exportConnection">{{ $t('message.export') }}</el-button>
+              <el-button icon="el-icon-download" @click="showImportDialog">{{ $t('message.import') }}</el-button>
+            </el-form-item>
+          </el-col>
+        </el-row>
+      </el-card>
 
-      <!-- keys per loading -->
-      <el-form-item>
-        <el-input-number
-          size="mini"
-          placeholder='500'
-          :min=10
-          :max=10000
-          :step=50
-          v-model='form.keysPageSize'>
-        </el-input-number>
-        <span slot="label">
-          {{ $t('message.keys_per_loading') }}
-          <el-popover
-            :content="$t('message.keys_per_loading_tip')"
-            placement="top-start"
-            trigger="hover">
-            <i slot="reference" class="el-icon-question"></i>
-          </el-popover>
-        </span>
-      </el-form-item>
-
-      <!-- export connections -->
-      <el-form-item :label="$t('message.config_connections')">
-        <el-button icon="el-icon-upload2" @click="exportConnection">{{ $t('message.export') }}</el-button>
-        <el-button icon="el-icon-download" @click="showImportDialog">{{ $t('message.import') }}</el-button>
-      </el-form-item>
-
-      <!-- font-family -->
-      <el-form-item :label="$t('message.font_family')">
-        <span slot="label">
-          {{ $t('message.font_family') }}
-          <el-popover
-            placement="top-start"
-            :title="$t('message.font_faq_title')"
-            trigger="hover">
-            <i slot="reference" class="el-icon-question"></i>
-            <p v-html="$t('message.font_faq')"></p>
-          </el-popover>
-          <i v-if="loadingFonts" class="el-icon-loading"></i>
-        </span>
-
-        <!-- font-family select -->
-        <el-select v-model="form.fontFamily" @visible-change="getAllFonts" allow-create default-first-option filterable multiple >
-          <el-option
-            v-for="(font, index) in allFonts"
-            :key="index"
-            :label="font"
-            :value="font">
-            <!-- for better performance, do not display font-family -->
-            <!-- :style="{'font-family': font}"> -->
-          </el-option>
-        </el-select>
-      </el-form-item>
-
-      <!-- current version -->
-      <el-form-item :label="$t('message.pre_version')" class='current-version'>
-        <el-tag type="info">{{ appVersion }}</el-tag>
-
-        <a href="###" @click.stop.prevent="showHotkeys">{{ $t('message.hotkey') }}</a>
-        <a href="###" @click.stop.prevent="clearCache">{{ $t('message.clear_cache') }}</a>
-        <a href="###" @click.stop.prevent="checkUpdate">{{ $t('message.check_update') }}</a>
-        <a href="https://github.com/qishibo/AnotherRedisDesktopManager/releases">{{ $t('message.manual_update') }}</a>
-        <a href="https://github.com/qishibo/AnotherRedisDesktopManager/">{{ $t('message.project_home') }}</a>
-      </el-form-item>
-
-      <!-- import file dialog -->
-      <el-dialog
-        width="400px"
-        :title="$t('message.select_import_file')"
-        :visible.sync="importConnectionVisible"
-        append-to-body>
-
-        <el-upload
-          ref="configUpload"
-          :auto-upload="false"
-          :multiple="false"
-          action=""
-          :limit="1"
-          :on-change="loadConnectionFile"
-          drag>
-          <i class="el-icon-upload"></i>
-          <div class="el-upload__text">{{ $t('message.put_file_here') }}</div>
-        </el-upload>
-
-        <div slot="footer" class="dialog-footer">
-          <el-button @click="importConnnection">{{ $t('el.messagebox.confirm') }}</el-button>
+      <el-card class="setting-card">
+        <div slot="header">
+          {{$t('message.pre_version')}}
+          <el-tag type="info">{{ appVersion }}</el-tag>
         </div>
-      </el-dialog>
+        <div class="current-version">
+          <a href="###" @click.stop.prevent="showHotkeys">{{ $t('message.hotkey') }}</a>
+          <a href="###" @click.stop.prevent="clearCache">{{ $t('message.clear_cache') }}</a>
+          <a href="###" @click.stop.prevent="checkUpdate">{{ $t('message.check_update') }}</a>
+          <a href="https://github.com/qishibo/AnotherRedisDesktopManager/releases">{{ $t('message.manual_update') }}</a>
+          <a href="https://github.com/qishibo/AnotherRedisDesktopManager/">{{ $t('message.project_home') }}</a>
+        </div>
+      </el-card>
     </el-form>
+
+    <!-- import file dialog -->
+    <el-dialog
+      width="400px"
+      :title="$t('message.select_import_file')"
+      :visible.sync="importConnectionVisible"
+      append-to-body>
+
+      <el-upload
+        ref="configUpload"
+        :auto-upload="false"
+        :multiple="false"
+        action=""
+        :limit="1"
+        :on-change="loadConnectionFile"
+        drag>
+        <i class="el-icon-upload"></i>
+        <div class="el-upload__text">{{ $t('message.put_file_here') }}</div>
+      </el-upload>
+
+      <div slot="footer" class="dialog-footer">
+        <el-button @click="importConnnection">{{ $t('el.messagebox.confirm') }}</el-button>
+      </div>
+    </el-dialog>
 
     <div slot="footer" class="dialog-footer">
       <el-button @click="visible = false">{{ $t('el.messagebox.cancel') }}</el-button>
       <el-button type="primary" @click="saveSettings">{{ $t('el.messagebox.confirm') }}</el-button>
     </div>
+
   </el-dialog>
 </template>
 
 <script type="text/javascript">
 import storage from '@/storage.js';
-import { ipcRenderer } from 'electron';
+import {ipcRenderer} from 'electron';
 import LanguageSelector from '@/components/LanguageSelector';
 
 export default {
   data() {
     return {
       visible: false,
-      form: {fontFamily: '', zoomFactor: 1.0, keysPageSize: 500},
+      form: {
+        fontFamily: '',
+        zoomFactor: 1.0,
+        keysPageSize: 500,
+        showLoadAllKeys: false,
+      },
       importConnectionVisible: false,
       connectionFileContent: '',
       appVersion: (new URL(window.location.href)).searchParams.get('version'),
@@ -143,7 +172,7 @@ export default {
       darkMode: localStorage.theme == 'dark',
     };
   },
-  components: { LanguageSelector },
+  components: {LanguageSelector},
   methods: {
     show() {
       this.visible = true;
@@ -164,7 +193,7 @@ export default {
     },
     changeZoom() {
       const {webFrame} = require('electron');
-      let zoomFactor   = this.form.zoomFactor;
+      let zoomFactor = this.form.zoomFactor;
 
       zoomFactor = zoomFactor ? zoomFactor : 1.0;
       webFrame.setZoomFactor(zoomFactor);
@@ -174,7 +203,9 @@ export default {
     },
     loadConnectionFile(file) {
       const reader = new FileReader();
-      reader.onload = (event) => { this.connectionFileContent = event.target.result; };
+      reader.onload = (event) => {
+        this.connectionFileContent = event.target.result;
+      };
       reader.readAsText(file.raw);
     },
     importConnnection() {
@@ -248,7 +279,8 @@ export default {
         localStorage.clear();
         this.$message.success(this.$t('message.delete_success'));
         window.location.reload();
-      }).catch(e => {});
+      }).catch(e => {
+      });
     },
     showHotkeys() {
       this.$parent.$refs.hotKeysDialog.show();
@@ -262,11 +294,48 @@ export default {
 </script>
 
 <style type="text/css">
-  .dark-mode .el-upload-dragger {
-    background: inherit;
-  }
-  .current-version a {
-    color: grey;
-    font-size: smaller;
-  }
+.setting-main-dialog {
+  width: 80%;
+  max-width: 900px;
+  margin-top: 7vh !important;
+}
+
+.dark-mode .el-upload-dragger {
+  background: inherit;
+}
+
+.setting-main-dialog .current-version a {
+  color: grey;
+  font-size: 95%;
+}
+
+.setting-main-dialog .setting-card {
+  margin-bottom: 8px;
+}
+.setting-main-dialog .setting-card .el-card__header {
+  font-size: 105%;
+  font-weight: bold;
+}
+
+.setting-main-dialog .setting-card .setting-row {
+  flex-wrap: wrap;
+}
+
+/* add height: fix el-select jitter when multiple*/
+.setting-main-dialog .setting-card .setting-row .setting-font-select .el-select__tags .el-tag {
+  height: 21px;
+  max-width: 98%;
+}
+
+/*label style inside el-select multiple*/
+.setting-main-dialog .setting-card .setting-row .setting-font-select .el-select__tags .el-tag .el-select__tags-text {
+  display: inline-block;
+  max-width: 90%;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+/*fix close icon vertical align*/
+.setting-main-dialog .setting-card .setting-row .setting-font-select .el-select__tags .el-tag .el-tag__close {
+  vertical-align: super;
+}
 </style>
