@@ -1,7 +1,7 @@
 <template>
   <div>
     <!-- </textarea> -->
-    <el-input ref='textInput' :disabled='disabled' type='textarea' :rows='textrows' :value='contentDisplay' @input='inputContent'>
+    <el-input ref='textInput' :disabled='disabled' type='textarea' v-model='contentDisplay' @input='inputContent'>
     </el-input>
   </div>
 </template>
@@ -11,20 +11,16 @@ export default {
   data() {
     return {
       confirmChange: false,
+      contentDisplay: "",
+      oldContentDisplay: "",
     };
   },
-  props: ['content', 'contentVisible', 'textrows', 'disabled'],
-  computed: {
-    contentDisplay() {
-      return this.content.toString();
-    },
-  },
+  props: ['content', 'contentVisible', 'disabled'],
   watch: {
-    content() {
+    content(val) {
       // refresh
-      this.$nextTick(() => {
-        this.$refs.textInput.$refs.textarea.value = this.contentDisplay;
-      });
+      this.contentDisplay = val.toString();
+      this.oldContentDisplay = this.contentDisplay;
     },
   },
   methods: {
@@ -34,8 +30,7 @@ export default {
         return this.content;
       }
 
-      const content = this.$refs.textInput.$refs.textarea.value;
-      return Buffer.from(content);
+      return Buffer.from(this.contentDisplay);
     },
     inputContent(value) {
       // visible content do nothing
@@ -52,9 +47,13 @@ export default {
         return this.confirmChange = true;
       }).catch(_ => {
         // recovery the input value
-        this.$refs.textInput.$refs.textarea.value = this.contentDisplay
+        this.contentDisplay = this.oldContentDisplay;
       });
     },
+  },
+  mounted() {
+    this.contentDisplay = this.content.toString();
+    this.oldContentDisplay = this.contentDisplay;
   },
 }
 </script>
