@@ -3,9 +3,11 @@ const it = {
     new_connection: 'Nuova Connessione',
     refresh_connection: 'Ricaricare',
     edit_connection: 'Modificare Connessione',
+    duplicate_connection: 'Copia connessione',
     del_connection: 'Elimina Connessione',
     close_connection: 'Chiudere Connessione',
     add_new_line: 'Inserisci Nuova Riga',
+    dump_to_clipboard: "Copia come comando",
     redis_version: 'Versione del Redis',
     process_id: 'ID del Processi',
     used_memory: 'Memoria Usata',
@@ -20,6 +22,9 @@ const it = {
     memory: 'Memoria',
     stats: 'Statistiche',
     settings: 'Impostazioni',
+    ui_settings: 'Aspetto esteriore',
+    feature_settings: 'Funzione',
+    common_settings: 'Generale',
     confirm_to_delete_row_data: 'Confermare per eliminare i dati della riga?',
     delete_success: 'Eliminato con successo',
     delete_failed: 'Eliminazione non riuscita',
@@ -93,7 +98,7 @@ const it = {
     private_key_faq: 'La chiave privata in formato RSA è compatibile e inizia con <pre>-----BEGIN RSA PRIVATE KEY-----</pre>\
     ma se inizi con<pre>-----BEGIN OPENSSH PRIVATE KEY-----</pre>è necessario convertire il formato tramite <pre>ssh-keygen -p -m pem -f ~/.ssh/id_rsa</pre>Questa operazione non influirà sul precedente accesso con chiave privata',
     dark_mode: 'Modalità scura',
-    load_more_keys: 'Carica più chiavi',
+    load_more_keys: 'carica di più',
     key_name: 'Nome della chiave',
     project_home: 'Progetto Home',
     cluster_faq: 'Seleziona qualsiasi nodo nel cluster da compilare e gli altri nodi verranno identificati automaticamente.',
@@ -145,6 +150,20 @@ const it = {
     hide_window: 'Nascondi finestra',
     minimize_window: 'Riduci finestra',
     maximize_window: 'Massimizza finestra',
+    load_all_keys: 'carica tutto',
+    show_load_all_keys: 'Abilita pulsante per caricare tutte le chiavi',
+    load_all_keys_tip: 'Carica tutte le chiavi contemporaneamente. Se il numero di chiavi è troppo grande, il client potrebbe rimanere bloccato. Si prega di usarlo correttamente',
+    tree_node_overflow: 'Troppi tasti o cartelle, tenere solo {num} per la visualizzazione. Se la tua chiave non è qui, si raccomanda la ricerca sfocata, o impostare il separatore per distribuire le chiavi in cartelle',
+    connection_readonly: 'Modalità di sola lettura. È vietato aggiungere, modificare ed eliminare',
+    memory_analysis: 'Analisi della memoria',
+    begin: 'Inizio',
+    pause: 'Pausa',
+    restart: 'Ricomincia',
+    max_display: 'Numero massimo di visualizzazioni: {num}',
+    max_scan: 'Numero massimo di scansioni: {num}',
+    close_left: 'Chiudi le schede a sinistra',
+    close_right: 'Chiudi le schede a destra',
+    close_other: 'Chiudi altre schede',
   },
 };
 
