@@ -4,8 +4,10 @@ const pt = {
     refresh_connection: 'Atualizar',
     edit_connection: 'Editar Conexão',
     del_connection: 'Deletar Conexão',
+    duplicate_connection: 'Copiar Conexão',
     close_connection: 'Encerrar Conexão',
     add_new_line: 'Adicionar nova linha',
+    dump_to_clipboard: "Copiar como comando",
     redis_version: 'Versão do Redis',
     process_id: 'ID do processo',
     used_memory: 'Memória utilizada',
@@ -20,6 +22,9 @@ const pt = {
     memory: 'Memória',
     stats: 'Estatísticas',
     settings: 'Configurações',
+    ui_settings: 'Aparência',
+    feature_settings: 'Função',
+    common_settings: 'Geral',
     confirm_to_delete_row_data: 'Você deseja excluir os dados da linha?',
     delete_success: 'Deleção executada com sucesso',
     delete_failed: 'Houve uma falha na deleção',
@@ -93,7 +98,7 @@ const pt = {
     private_key_faq: 'A chave privada de formato RSA é compatível e começa com <pre>-----BEGIN RSA PRIVATE KEY-----</pre>\
     mas se começar com<pre>-----BEGIN OPENSSH PRIVATE KEY-----</pre>você precisa converter o formato via <pre>ssh-keygen -p -m pem -f ~/.ssh/id_rsa</pre>Esta operação não afetará o login de chave privada anterior',
     dark_mode: 'Modo noturno',
-    load_more_keys: 'Carregar mais chaves',
+    load_more_keys: 'carregar mais',
     key_name: 'Nome da Chave',
     project_home: 'Home do Projeto',
     cluster_faq: 'Selecione qualquer nó no cluster para preencher e outros nós serão identificados automaticamente.',
@@ -145,6 +150,20 @@ const pt = {
     hide_window: 'Ocultar janela',
     minimize_window: 'Minimize a janela',
     maximize_window: 'Maximize a janela',
+    load_all_keys: 'carregar tudo',
+    show_load_all_keys: 'Habilite o botão para carregar todas as chaves',
+    load_all_keys_tip: 'Carregue todas as chaves de uma vez. Se o número de chaves for muito grande, o cliente pode ficar preso. Por favor, use-o corretamente',
+    tree_node_overflow: 'Muitas teclas ou pastas, manter apenas {num} para exibição. Se a sua chave não está aqui, é recomendada a pesquisa difusa, ou configurar o separador para espalhar as chaves em pastas',
+    connection_readonly: 'Modo somente leitura. Adição, edição e exclusão são proibidas',
+    memory_analysis: 'Análise de memória',
+    begin: 'Começar',
+    pause: 'Pausa',
+    restart: 'Reiniciar',
+    max_display: 'Número máximo de exibições: {num}',
+    max_scan: 'Número máximo de verificações: {num}',
+    close_left: 'Fechar abas esquerdas',
+    close_right: 'Fechar abas direitas',
+    close_other: 'Fechar outras guias',
   },
 };
 
