@@ -3,13 +3,15 @@ const ru = {
     new_connection: 'Новое подключение',
     refresh_connection: 'Обновить',
     edit_connection: 'Редактировать подключение',
+    duplicate_connection: 'Копировать соединение',
     del_connection: 'Удалить подключение',
     close_connection: 'Закрыть соединение',
     add_new_line: 'Добавить новую строку',
+    dump_to_clipboard: "Копировать как команду",
     redis_version: 'Версия Redis',
     process_id: 'ID процесса',
     used_memory: 'Использовано памяти',
-    used_memory_peak: 'Максимальный объем памяти',
+    used_memory_peak: 'Пиковый объем памяти',
     used_memory_lua: 'Использование памяти Lua',
     connected_clients: 'Подключено клиентов',
     total_connections_received: 'Количество подключений',
@@ -20,6 +22,9 @@ const ru = {
     memory: 'Оперативная память',
     stats: 'Статистика',
     settings: 'Настройки',
+    ui_settings: 'Внешность',
+    feature_settings: 'Функция',
+    common_settings: 'Общий',
     confirm_to_delete_row_data: 'Подтвердить удаление данной строки?',
     delete_success: 'Успешно удалено',
     delete_failed: 'Удалить не удалось',
@@ -44,7 +49,7 @@ const ru = {
     php_unserialize_format_failed: 'PHP Unserialize форматирование не удалось',
     clean_up: 'Очистить',
     redis_console: 'Redis console',
-    confirm_to_delete_connection: 'Вы уверены, что хотите удалили подключение?',
+    confirm_to_delete_connection: 'Вы уверены, что хотите удалить подключение?',
     connection_exists: 'Настройка с таким подключением уже существует',
     close_to_edit_connection: 'Вы должны закрыть соединение перед редактированием. Вы уверены, что хотите продолжить?',
     close_to_connection: 'Подтверждение закрытия соединения?',
@@ -108,7 +113,7 @@ const ru = {
     delete_folder: 'Сканировать и удалить всю папку',
     multiple_select: 'Множественный выбор',
     copy: 'Копировать',
-    copy_success: 'Копировать успешно',
+    copy_success: 'Успешно скопировано',
     keys_to_be_deleted: 'Ключи для удаления',
     delete_all: 'Удалить все',
     clear_cache: 'Oчистить кэш',
@@ -117,7 +122,7 @@ const ru = {
     toggle_check_all: 'выбрать все | отменить выбор',
     select_lang: 'Выбрать язык',
     clear_cache_tip: 'Когда возникает проблема с клиентом, эта операция удалит все подключения и конфигурации для восстановления клиента',
-    detail: 'Деталь',
+    detail: 'Информация',
     separator_tip: 'Разделитель древовидного представления, установлен на пустой, чтобы отключить дерево и отображать ключи в виде списка',
     confirm_modify_unvisible_content: 'Контент содержит невидимые символы, вы можете смело редактировать его в «Hex View». Если продолжение редактирования в «Text View» может вызвать ошибки кодирования, обязательно продолжить?',
     keys_per_loading: 'Количество ключей',
@@ -126,7 +131,7 @@ const ru = {
     port: 'Порт',
     username: 'Имя пользователя',
     password: 'Пароль',
-    connection_name: 'Произвольное имя',
+    connection_name: 'Имя соединения',
     separator: 'Pазделитель',
     timeout: 'Tайм-аут',
     private_key: 'Закрытый ключ',
@@ -134,17 +139,31 @@ const ru = {
     authority: 'Авторизация',
     redis_node_password: 'Пароль узла Redis',
     master_group_name: 'Master Имя группы',
-    command_log: 'Бревно',
+    command_log: 'Журнал логов',
     sentinel_faq: 'Вы можете выбрать один из нескольких дозорных. Пожалуйста, заполните конфигурацию дозорных для адреса, порта и пароля. Пароль узла Redis - это пароль главного узла, который контролируется дозорным.',
-    hotkey: 'Горячая клавиша',
+    hotkey: 'Горячие клавиши',
     persist: 'Удалить время истечения',
     custom_formatter: 'Пользовательский форматировщик',
     edit: 'Редактировать',
-    new: 'Добавлять',
+    new: 'Добавить',
     custom: 'Настроить',
     hide_window: 'Скрыть окно',
     minimize_window: 'Свернуть окно',
     maximize_window: 'Развернуть окно',
+    load_all_keys: 'загрузить все',
+    show_load_all_keys: 'Кнопка включения для загрузки всех ключей',
+    load_all_keys_tip: 'Загрузите все ключи одновременно. Если количество ключей слишком велико, клиент может застрять. Пожалуйста, используйте его правильно',
+    tree_node_overflow: 'Слишком много клавиш или папок, оставьте только {num} для отображения. Если вашего ключа здесь нет, рекомендуется использовать нечеткий поиск, или Настройка разделителя для разделения ключей в папку',
+    connection_readonly: 'Режим только для чтения. Добавление, редактирование и удаление запрещено',
+    memory_analysis: 'Анализ памяти',
+    begin: 'Начинать',
+    pause: 'Пауза',
+    restart: 'Начать сначала',
+    max_display: 'Максимальное количество дисплеев: {num}',
+    max_scan: 'Максимальное количество сканирований: {num}',
+    close_left: 'Закрыть левые вкладки',
+    close_right: 'Закрыть правые вкладки',
+    close_other: 'Закрыть другие вкладки',
   },
 };
 
