@@ -3,9 +3,11 @@ const tr = {
     new_connection: 'Yeni Bağlantı',
     refresh_connection: 'Yenile',
     edit_connection: 'Bağlantıyı Düzenle',
+    duplicate_connection: 'Bağlantıyı Kopyala',
     del_connection: 'Bağlantıyı Sil',
     close_connection: 'Bağlantıyı Kapat',
     add_new_line: 'Yeni Satır Ekle',
+    dump_to_clipboard: "Komut olarak kopyala",
     redis_version: 'Redis Sürümü',
     process_id: 'Süreç Kimliği',
     used_memory: 'Kullanılan Bellek',
@@ -20,6 +22,9 @@ const tr = {
     memory: 'Bellek',
     stats: 'İstatistik',
     settings: 'Ayarlar',
+    ui_settings: 'Dış görünüş',
+    feature_settings: 'Fonksiyon',
+    common_settings: 'Genel',
     confirm_to_delete_row_data: 'Satır Verilerini Silmek İstiyor Musunuz?',
     delete_success: 'Silme Başarılı',
     delete_failed: 'Silme Başarısız',
@@ -54,7 +59,7 @@ const tr = {
     enter_new_key: 'Önce Yeni Anahtarın Adını Girin',
     key_type: 'Anahtar Tipi',
     save: 'Kaydet',
-    enter_to_search: 'Aramak Yapmak İçin',
+    enter_to_search: 'Arama Yapmak İçin',
     export_success: 'Dışa Aktarma Başarılı',
     select_import_file: 'Dosyayı Seçin',
     import_success: 'İçe Aktarma Başarılı',
@@ -108,7 +113,7 @@ const tr = {
     delete_folder: 'Tüm Klasörü Tara ve Sil',
     multiple_select: 'Çoklu Seçim',
     copy: 'Kopyala',
-    copy_success: 'Başarıyla kopyalayın',
+    copy_success: 'Başarıyla kopyalandı',
     keys_to_be_deleted: 'Silinecek Anahtarlar',
     delete_all: 'Hepsini sil',
     clear_cache: 'Önbelleği Temizle',
@@ -123,10 +128,10 @@ const tr = {
     keys_per_loading: 'Anahtar Sayısı',
     keys_per_loading_tip: 'Her seferinde yüklenen anahtar sayısı. Çok büyük ayarlamak performansı etkileyebilir',
     host: 'Adres',
-    port: 'Liman',
+    port: 'Port',
     username: 'Kullanıcı adı',
     password: 'Parola',
-    connection_name: 'Özel ad',
+    connection_name: 'Bağlantı adı',
     separator: 'Ayırıcı',
     timeout: 'Zaman aşımı',
     private_key: 'Özel anahtar',
@@ -141,10 +146,24 @@ const tr = {
     custom_formatter: 'Özel Biçimlendirici',
     edit: 'Düzenle',
     new: 'Ekle',
-    custom: 'Özelleştirmek',
+    custom: 'Özelleştir',
     hide_window: 'Pencereyi gizle',
     minimize_window: 'Pencereleri küçültür',
     maximize_window: 'Pencereyi büyüt',
+    load_all_keys: 'hepsini yükle',
+    show_load_all_keys: 'Tüm anahtarları yüklemek için etkinleştir düğmesi',
+    load_all_keys_tip: 'Tüm anahtarları bir kerede yükleyin. Anahtar sayısı çok fazlaysa, istemci takılabilir. Lütfen doğru kullanın',
+    tree_node_overflow: 'Çok fazla anahtar veya klasör var, yalnızca {num} görüntüleme için ayrıldı. Anahtarınız burada değilse bulanık arama önerilir, veya anahtarı klasöre yaymak için bir ayırıcı ayarlayın',
+    connection_readonly: 'Salt okunur mod. Eklemek, düzenlemek ve silmek yasaktır',
+    memory_analysis: 'Bellek Analizi',
+    begin: 'Başlat',
+    pause: 'Duraklat',
+    restart: 'Tekrar başlat',
+    max_display: 'Maksimum ekran sayısı: {num}',
+    max_scan: 'Maksimum tarama sayısı: {num}',
+    close_left: 'Sol Sekmeleri Kapat',
+    close_right: 'Sağ Sekmeleri Kapat',
+    close_other: 'Diğer Sekmeleri Kapat',
   },
 };
 
